refactor(devtools): split pipeline_probe monolith, add Click dispatch (#625)
Split the 1484-line devtools/pipeline_probe.py into a 6-module package
(engine, request, result, report, staging, __init__) with backward-compatible
imports. Replaced the argparse dispatch in __main__ with a Click-based group
that auto-generates commands from the CommandSpec catalog. The Click dispatch
preserves --json forwarding and --list-commands --json machine output.
